Add files via upload
Adjustments regarding:
i. Reference period for individual-level regression.
ii. Restructured the regressions on the number of students and schools in the SAEB data.
iii. Saving plots directly in the Report LaTeX folder for improved efficiency.
iv. Added a separate dataframe for the age DiD specification (differs from the data used to access the student age result).