Add files via upload
2. data_desc.R
i. Changed the color scale for better visualization of municipalities that did not participate in the FUNDEF.

3. regression_v2.R
i. Added regression with the total number of employees.

3. regression_v3.R
i. New version of the regression codes.
ii. Focus only on the aluno_dosage variable.
iii. Removed the spending analysis for Elementary and High School. Better focus on Preschool and Daycare.
iv. Still under improvement.